Updated index to match other pages' code format
